Move useDatabaseInfoMaps into its own hook file.
