+++ src/shared/contexts/databaseInfoMaps.tsx
@@ -1,4 +1,4 @@
-import { createContext, useContext, ReactNode, useMemo } from 'react';
+import { createContext, ReactNode, useMemo } from 'react';
 import { DatabaseInfo } from '../../uniprotkb/types/databaseRefs';
 import {
   DatabaseInfoMaps,
@@ -6,9 +6,10 @@ import {
 } from '../../uniprotkb/utils/database';
 import apiUrls from '../config/apiUrls';
 import useDataApi from '../hooks/useDataApi';
-import * as logging from '../utils/logging';
 
-const DatabaseInfoMapsContext = createContext<DatabaseInfoMaps | null>(null);
+export const DatabaseInfoMapsContext = createContext<DatabaseInfoMaps | null>(
+  null
+);
 
 type DatabaseInfoMapsProviderProps = {
   children: ReactNode;
@@ -29,14 +30,4 @@ export const DatabaseInfoMapsProvider = ({
       {children}
     </DatabaseInfoMapsContext.Provider>
   ) : null;
-};
-
-export const useDatabaseInfoMaps = () => {
-  const context = useContext(DatabaseInfoMapsContext);
-  if (context === undefined) {
-    logging.error(
-      'useDatabaseInfoMaps must be used within a DatabaseInfoMapsProvider'
-    );
-  }
-  return context;
 };
